added float functionality to printf

+++ lab11/analog18f.c
@@ -58,8 +58,8 @@ double analogMap(
             unsigned int analog_reading,
             unsigned int analog_max,
             unsigned int analog_min,
-            double sensor_max,
-            double sensor_min)
+            float sensor_max,
+            float sensor_min)
     {
-    return (sensor_min + ((analog_reading-analog_min)/((double)(analog_max-analog_min)))*(sensor_max-sensor_min));
+    return (sensor_min + ((analog_reading-analog_min)/((float)(analog_max-analog_min)))*(sensor_max-sensor_min));
     }
